fix: fixed the PR review issues
