+++ backend/src/modules/alert/alert.service.ts
@@ -14,26 +14,7 @@ import { Outcome } from 'src/utils/types/outcome';
 import { EventLogService } from '../event_log/eventLog.service';
 import { GoldLakehouseService } from '../gold-lakehouse/gold-lakehouse.service';
 import { transactionDataResponseDTO } from './dto/transactionHistory.dto';
-
-interface RuleResult {
-  id: string;
-  cfg: string;
-  wght: number;
-  prcgTm: number;
-  tenantId: string;
-  subRuleRef: string;
-  indpdntVarbl: number;
-}
-
-interface AlertedTypology {
-  id: string;
-  cfg: string;
-  result: number;
-  alertThreshold: number;
-  interdictionThreshold: number;
-  ruleResults: RuleResult[];
-}
-
+import {AlertedTypology} from './types/alert.types'
 @Injectable()
 export class AlertService {
   constructor(
