+++ backend/src/modules/alert/types/alert.types.ts
@@ -1,0 +1,18 @@
+interface RuleResult {
+  id: string;
+  cfg: string;
+  wght: number;
+  prcgTm: number;
+  tenantId: string;
+  subRuleRef: string;
+  indpdntVarbl: number;
+}
+
+export interface AlertedTypology {
+  id: string;
+  cfg: string;
+  result: number;
+  alertThreshold: number;
+  interdictionThreshold: number;
+  ruleResults: RuleResult[];
+}
