+++ frontend/src/features/alerts/pages/AlertsDashboard.tsx
@@ -170,18 +170,7 @@ const AlertsDashboard: React.FC = () => {
         return 'text-gray-600 bg-gray-50';
     }
   };
-
-  // const getStatusColor = (status?: string): string => {
-  //   switch (status) {
-  //     case 'ALRT':
-  //       return 'text-red-700 bg-red-50 ring-1 ring-red-200';
-  //     case 'NALT':
-  //       return 'text-green-700 bg-green-50 ring-1 ring-green-200';
-  //     default:
-  //       return 'text-gray-700 bg-gray-50 ring-1 ring-gray-200';
-  //   }
-  // };
-
+  
   const columns: Array<AlertsTableColumn<Alert>> = [
     {
       key: 'alert_id',
@@ -228,24 +217,6 @@ const AlertsDashboard: React.FC = () => {
       header: 'Source',
       sortable: true,
     },
-    // {
-    //   key: 'alert_status',
-    //   header: 'Status',
-    //   sortable: false,
-    //   render: (_, alert) => {
-    //     const status = alert.alert_data?.status as string | undefined;
-
-    //     return (
-    //       <span
-    //         className={`inline-flex px-2 py-1 text-xs font-medium rounded-md ${getStatusColor(
-    //           status,
-    //         )}`}
-    //       >
-    //         {status || 'N/A'}
-    //       </span>
-    //     );
-    //   },
-    // },
     {
       key: 'alert_type',
       header: 'Alert Type',
